test(e2e): share mockauth sign-in
