+++ test/e2e/fixtures.ts
@@ -1,29 +1,10 @@
 import type { Page } from '@playwright/test';
 import { test as base, expect } from '@playwright/test';
+import { signInViaMockAuth } from './sign-in-helper';
 export { expect };
 
-const defaultEmail = 'e2e-tester@agyn.test';
-const expectedEmail = process.env.E2E_OIDC_EMAIL ?? defaultEmail;
-
 async function signInAndLoad(page: Page) {
-  await page.goto('/');
-
-  await page.waitForURL(/mockauth\.dev\/r\/.*\/oidc\/login/);
-
-  const strategyTabs = page.getByTestId('login-strategy-tabs');
-  if (await strategyTabs.isVisible()) {
-    await strategyTabs.getByRole('tab', { name: 'Email' }).click();
-  }
-
-  const emailInput = page.getByTestId('login-email-input');
-  await expect(emailInput).toBeVisible();
-  await emailInput.fill(expectedEmail);
-
-  await page.getByRole('button', { name: 'Continue' }).click();
-
-  await page.waitForURL(/\/agents\/threads/);
-  const threadsList = page.getByTestId('threads-list');
-  await expect(threadsList).toBeVisible();
+  await signInViaMockAuth(page);
 }
 
 export const test = base.extend({
