+++ test/e2e/sign-in-helper.ts
@@ -1,0 +1,39 @@
+import type { Page } from '@playwright/test';
+import { expect } from '@playwright/test';
+
+const defaultEmail = 'e2e-tester@agyn.test';
+
+type SignInOptions = {
+  onLoginPage?: (page: Page) => Promise<void>;
+};
+
+export async function signInViaMockAuth(
+  page: Page,
+  email?: string,
+  options: SignInOptions = {},
+) {
+  const expectedEmail = email ?? process.env.E2E_OIDC_EMAIL ?? defaultEmail;
+
+  await page.goto('/');
+
+  await page.waitForURL(/mockauth\.dev\/r\/.*\/oidc\/login/);
+
+  if (options.onLoginPage) {
+    await options.onLoginPage(page);
+  }
+
+  const strategyTabs = page.getByTestId('login-strategy-tabs');
+  if (await strategyTabs.isVisible()) {
+    await strategyTabs.getByRole('tab', { name: 'Email' }).click();
+  }
+
+  const emailInput = page.getByTestId('login-email-input');
+  await expect(emailInput).toBeVisible();
+  await emailInput.fill(expectedEmail);
+
+  await page.getByRole('button', { name: 'Continue' }).click();
+
+  await page.waitForURL(/\/agents\/threads/);
+  const threadsList = page.getByTestId('threads-list');
+  await expect(threadsList).toBeVisible();
+}
